Remove last modified date display
Because it refers to the last modified date for the line,
not for the actual status.

While is it possible to get the date for the particular status,
seems like that date is not correct so better not to display it,
given also that often you can find information regarding the affected days
directly on the disruption description.